refactor : constants 상수값 정리
ui 영역인 view와 babel 플러그인에서 사용되는 상수값을 분리
나머지 상수값은 index.js를 통해 관리
기존 불필요하게 흩어져있는 상수값을 객체에 넣어서 한번에 관리하도록 수정
