+++ bin/utils/findFiles.js
@@ -1,6 +1,8 @@
 import fs from 'fs';
 import path from 'path';
-import {PATH} from "../../constants.js";
+
+
+import {PATH} from "../../constants/index.js";
 
 /**
  * 테스트 파일을 찾는 로직입니다.
