+++ constants/babel.js
@@ -1,0 +1,12 @@
+export const MOCK = {
+  STORE_NAME: 'mockStore',
+  STORE_PATH: '@dannysir/js-te/src/mock/store.js'
+};
+
+export const BABEL = {
+  MODULE: 'module',
+  CONST: 'const',
+  HAS: 'has',
+  GET: 'get',
+  PERIOD: '.',
+};
