+++ constants/index.js
@@ -1,0 +1,20 @@
+export const PATH = {
+  NODE_MODULES: 'node_modules',
+  TEST_DIRECTORY: 'test',
+  TEST_FILE: '.test.js',
+  JAVASCRIPT_FILE: '.js',
+  BIN: 'bin',
+  DIST: 'dist',
+  DANNYSIR_JS_TE: '@dannysir/js-te',
+};
+
+export const RESULT_MSG = {
+  TESTS: 'Tests: ',
+  TOTAL: 'Total Result: ',
+  CHECK: '✓ ',
+  CROSS: '✗ ',
+  DIRECTORY_DELIMITER: ' > ',
+  EMPTY: '',
+};
+
+export const DEFAULT_COUNT = 0;
