+++ src/mock/collectMocks.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import {transformSync} from "@babel/core";
-import {BABEL} from "../../constants.js";
 import {createMockCollectorPlugin} from "../../babelPlugins/babelCollectMocks.js";
+import {BABEL} from "../../constants/babel.js";
 
 export const collectMockedPaths = (testFiles) => {
   const mockedPaths = new Set();
