+++ utils/formatString.js
@@ -1,14 +1,14 @@
-import {CHECK, CROSS, DIRECTORY_DELIMITER, EMPTY} from "../constants.js";
 import {green, red} from "./consoleColor.js";
+import {RESULT_MSG} from "../constants/index.js";
 
 export const formatSuccessMessage = (test) => {
-  const pathString = test.path === '' ? EMPTY : test.path + DIRECTORY_DELIMITER;
-  return green(CHECK) + pathString + test.description;
+  const pathString = test.path === '' ? RESULT_MSG.EMPTY : test.path + RESULT_MSG.DIRECTORY_DELIMITER;
+  return green(RESULT_MSG.CHECK) + pathString + test.description;
 };
 
 export const formatFailureMessage = (test, error) => {
   const messages = [];
-  messages.push(red(CROSS) + test.path + test.description);
+  messages.push(red(RESULT_MSG.CROSS) + test.path + test.description);
   messages.push(red(`  Error Message : ${error.message}`));
   return messages.join('\n');
 };
